Datafix.FrameworkBuilder

Description

Builds a DataFlowFramework for a Denotational formulation in terms of MonadDatafix. Effectively reduces descriptions from Datafix.Denotational to ones from Datafix.Explicit, so that solvers such as Datafix.Worklist only have to provide an interpreter for MonadDependency.

Documentation

data FrameworkBuilder m a

Constructs a build plan for a DataFlowFramework by tracking allocation of Nodes mapping to ChangeDetectors and transfer functions.

buildFramework :: forall m a. MonadDependency m => (forall md. (MonadDatafix md, DepM md ~ m) => md a) -> (a, Node, DataFlowFramework m)

(root, max, dff) = buildFramework builder executes the build plan specified by builder and returns the resulting DataFlowFramework dff, as well as the root Node denoting the transfer function returned by the FrameworkBuilder action and the maximum node of the problem as a proof for its denseness.