Datafix.NodeAllocator

Description

Helpers for allocating Nodes in an ergonomic manner, e.g. taking care to get mfix right under the hood for allocation in recursive bindings groups through the key primitive allocateNode.

Documentation

data NodeAllocator v a

A state monad wrapping a mapping from Node to some v which we will instantiate to appropriate LiftedFuncs.

allocateNode :: (Node -> NodeAllocator v (a, v)) -> NodeAllocator v a

Allocates the next Node, which is greater than any nodes requested before.

The value stored at that node is the result of a NodeAllocator computation which may already access the Node associated with that value. This is important for the case of recursive let, where the denotation of an expression depends on itself.

runAllocator :: NodeAllocator v a -> (a, Array v)

Runs the allocator, beginning with an empty mapping.