Datafix.Utils.TypeLevel

Description

Some type-level helpers for 'curry'/'uncurry'ing arbitrary function types.

Documentation

type family All (p :: k -> Constraint) (as :: [k]) :: Constraint where ...

All p as ensures that the constraint p is satisfied by all the types in as. (Types is between scare-quotes here because the code is actually kind polymorphic)

Equations

All p '[] = ()

All p (a ': as) = (p a, All p as)

type family If (b :: Bool) (l :: k) (r :: k) :: k where ...

On Booleans

Equations

If True l r = l

If False l r = r

type family Foldr (c :: k -> l -> l) (n :: l) (as :: [k]) :: l where ...

On Lists

Equations

Foldr c n '[] = n

Foldr c n (a ': as) = c a (Foldr c n as)

type family Foldr' (c :: Function k (Function l l -> *) -> *) (n :: l) (as :: [k]) :: l where ...

Version of Foldr taking a defunctionalised argument so that we can use partially applied functions.

Equations

Foldr' c n '[] = n

Foldr' c n (a ': as) = Apply (Apply c a) (Foldr' c n as)

type family Map (f :: Function k l -> *) (as :: [k]) :: [l] where ...

Equations

Map f as = Foldr' (ConsMap0 f) '[] as

data ConsMap0 :: (Function k l -> *) -> Function k (Function [l] [l] -> *) -> *

data ConsMap1 :: (Function k l -> *) -> k -> Function [l] [l] -> *

type family Constant (b :: l) (as :: [k]) :: [l] where ...

Equations

Constant b as = Map (Constant1 b) as

type Arrows (as :: [*]) (r :: *) = Foldr (->) r as

Arrows [a1,..,an] r corresponds to a1 -> .. -> an -> r

arrowsAxiom :: Arrows (ParamTypes func) (ReturnType func) :~: func

type family Products (as :: [*]) where ...

Products [] corresponds to (), Products [a] corresponds to a, Products [a1,..,an] corresponds to (a1, (..,( an)..)).

So, not quite a right fold, because we want to optimize for the empty, singleton and pair case.

Equations

Products '[] = ()

Products '[a] = a

Products (a ': as) = (a, Products as)

type family IsBase (t :: *) :: Bool where ...

IsBase t is 'True whenever t is *not* a function space.

Equations

IsBase (a -> t) = False

IsBase a = True

type family ParamTypes (t :: *) :: [*] where ...

Using IsBase we can define notions of ParamTypes and ReturnTypes which *reduce* under positive information IsBase t ~ 'True even though the shape of t is not formally exposed

Equations

ParamTypes t = If (IsBase t) '[] (ParamTypes' t)

type family ParamTypes' (t :: *) :: [*] where ...

Equations

ParamTypes' (a -> t) = a ': ParamTypes t

type family ReturnType (t :: *) :: * where ...

Equations

ReturnType t = If (IsBase t) t (ReturnType' t)

type family ReturnType' (t :: *) :: * where ...

Equations

ReturnType' (a -> t) = ReturnType t

class Currying as b where

Currying as b witnesses the isomorphism between Arrows as b and Products as -> b. It is defined as a type class rather than by recursion on a singleton for as so all of that these conversions are inlined at compile time for concrete arguments.

Minimal complete definition

uncurrys, currys

Methods

uncurrys :: Arrows as b -> Products as -> b

currys :: (Products as -> b) -> Arrows as b

data Function :: * -> * -> *

data Constant0 :: Function a (Function b a -> *) -> *

data Constant1 :: * -> Function b a -> *

type family Apply (t :: Function k l -> *) (u :: k) :: l