Datafix.Worklist.Graph

Description

Abstracts over the representation of the data-flow graph.

The contents of this module are more or less internal to the Datafix.Worklist implementation.

Documentation

data PointInfo domain

The data associated with each point in the transfer function of a data-flow Node.

Constructors

PointInfo

Fields

value :: !(Maybe (ReturnType domain))

The value at this point. Can be Nothing only when a loop was detected.

references :: !(IntArgsMonoSet (Products (ParamTypes domain)))

Points this point of the transfer function depends on.

referrers :: !(IntArgsMonoSet (Products (ParamTypes domain)))

Points depending on this point.

iterations :: !Int

The number of times this point has been updated through calls to updateNodeValue.

emptyPointInfo :: PointInfo domain

The default PointInfo.

data Diff a

Diff between two IntArgsMonoSets.

Constructors

Diff

Fields

added :: !(IntArgsMonoSet a)

removed :: !(IntArgsMonoSet a)

computeDiff :: MonoMapKey k => IntArgsMonoSet k -> IntArgsMonoSet k -> Diff k

Computes the diff between two IntArgsMonoSets.

class GraphRef (ref :: * -> *) where

Abstracts over the concrete representation of the data-flow graph.

There are two instances: The default Ref for sparse graphs based on an IntMap and Ref for the dense case, storing the Node mapping in a IOVector.

Minimal complete definition

updatePoint, lookup, lookupLT

Methods

updatePoint :: MonoMapKey (Products (ParamTypes domain)) => Int -> Products (ParamTypes domain) -> ReturnType domain -> IntArgsMonoSet (Products (ParamTypes domain)) -> ReaderT (ref domain) IO (PointInfo domain)

lookup :: MonoMapKey (Products (ParamTypes domain)) => Int -> Products (ParamTypes domain) -> ReaderT (ref domain) IO (Maybe (PointInfo domain))

lookupLT :: MonoMapKey (Products (ParamTypes domain)) => Int -> Products (ParamTypes domain) -> ReaderT (ref domain) IO [(Products (ParamTypes domain), PointInfo domain)]